Create play modal component
