+++ src/pages/film-room/[game]/[time]/index.tsx
@@ -1,62 +1,26 @@
 import ExpandLessIcon from "@mui/icons-material/ExpandLess";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import {
-  Button,
-  Checkbox,
-  FormControlLabel,
-  TextField,
-  Typography,
-} from "@mui/material";
+import { Button, Typography } from "@mui/material";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import Youtube, { type YouTubeEvent, type YouTubePlayer } from "react-youtube";
 import PlayDirectory from "~/components/play-directory";
-import Mentions from "~/components/play-mentions";
-import { useAuthContext } from "~/contexts/auth";
+import PlayModal from "~/components/play-modal";
 import { useMobileContext } from "~/contexts/mobile";
-import { useIsDarkContext } from "~/pages/_app";
 import { supabase } from "~/utils/supabase";
-import type { GameListType, PlayerType } from "~/utils/types";
-
-type PlayType = {
-  start: number | null | undefined;
-  end: number | null | undefined;
-};
-
-type NoteType = {
-  title: string;
-  note: string;
-  highlight: boolean;
-};
+import type { GameListType } from "~/utils/types";
 
 const FilmRoom = () => {
   const router = useRouter();
   const { screenWidth } = useMobileContext();
-  const { borderStyle } = useIsDarkContext();
-  const { user } = useAuthContext();
-  const [game, setGame] = useState<GameListType | null>(null);
-  const [isPlayDirectoryOpen, setIsPlayDirectoryOpen] =
-    useState<boolean>(false);
 
+  const [game, setGame] = useState<GameListType | null>(null);
   const [player, setPlayer] = useState<YouTubePlayer | null>(null);
 
-  const [affiliatedPlayers, setAffiliatedPlayers] = useState<PlayerType | null>(
-    null,
-  );
-  const [mentions, setMentions] = useState<string[]>([]);
+  const [isPlayModalOpen, setIsPlayModalOpen] = useState<boolean>(false);
 
-  const [isClipStarted, setIsClipStarted] = useState(false);
-  const [play, setPlay] = useState<PlayType>({
-    start: null,
-    end: null,
-  });
-  const [noteOpen, setNoteOpen] = useState<boolean>(false);
-  const [noteDetails, setNoteDetails] = useState<NoteType>({
-    title: "",
-    note: "",
-    highlight: false,
-  });
-  const [isValidPlay, setIsValidPlay] = useState<boolean>(false);
+  const [isPlayDirectoryOpen, setIsPlayDirectoryOpen] =
+    useState<boolean>(false);
 
   const fetchGame = async () => {
     const { data } = await supabase
@@ -69,29 +33,6 @@ const FilmRoom = () => {
     if (data) setGame(data);
   };
 
-  const fetchPlayers = async () => {
-    const { data } = await supabase
-      .from("affiliations")
-      .select(`user_id, profiles (name)`)
-      .neq("user_id", user.userId)
-      .match({ team_id: user.currentAffiliation?.team.id, role: "player" });
-    if (data) setAffiliatedPlayers(data);
-  };
-
-  const handleInput = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setNoteDetails({
-      ...noteDetails,
-      [name]: value,
-    });
-  };
-
-  const getCurrentTime = () => {
-    if (player) {
-      return player.getCurrentTime();
-    }
-  };
-
   const videoOnReady = (e: YouTubeEvent) => {
     const video = e.target;
     setPlayer(video);
@@ -101,101 +42,8 @@ const FilmRoom = () => {
     }
   };
 
-  const startClip = async () => {
-    setIsClipStarted(true);
-    const time = await player?.getCurrentTime();
-    const roundedTime = Math.round(time!);
-    setPlay({ ...play, start: roundedTime });
-    void player?.playVideo();
-  };
-
-  const endClip = async () => {
-    setIsClipStarted(false);
-    const time = await getCurrentTime();
-    const roundedTime = Math.round(time!);
-    void player?.pauseVideo();
-    setPlay({ ...play, end: roundedTime });
-    setNoteOpen(true);
-  };
-
-  const resetPlay = async () => {
-    setNoteOpen(false);
-    setNoteDetails({
-      title: "",
-      note: "",
-      highlight: false,
-    });
-    setPlay({ end: null, start: null });
-    setMentions([]);
-  };
-
-  const handleMention = async (player: string, play: string) => {
-    await supabase
-      .from("play_mentions")
-      .insert({
-        play_id: play,
-        sender_id: `${user.userId}`,
-        receiver_id: player,
-      })
-      .select()
-      .single();
-  };
-
-  const createPlay = async () => {
-    const { data } = await supabase
-      .from("plays")
-      .insert({
-        team_id: user.currentAffiliation?.team.id,
-        profile_id: user.userId,
-        game_id: `${game?.id}`,
-        highlight: noteDetails.highlight,
-        title: noteDetails.title,
-        note: noteDetails.note,
-        start_time: play.start!,
-        end_time: play.end!,
-        author_role: `${user.currentAffiliation?.role}`,
-        author_name: `${user.name}`,
-      })
-      .select()
-      .single();
-    if (data) {
-      mentions.forEach((mention) => {
-        const player = affiliatedPlayers?.find(
-          (v) => v.profiles?.name === mention,
-        );
-        if (player) {
-          void handleMention(player.user_id, data.id);
-        }
-      });
-      void resetPlay();
-    }
-  };
-
-  const checkValidForm = () => {
-    if (
-      typeof play.end !== "number" ||
-      typeof play.start !== "number" ||
-      noteDetails.note === "" ||
-      noteDetails.title === ""
-    ) {
-      setIsValidPlay(false);
-    } else {
-      setIsValidPlay(true);
-    }
-  };
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    void createPlay();
-  };
-
-  useEffect(() => {
-    checkValidForm();
-  }, [noteDetails, play]);
-
   useEffect(() => {
     if (router.query.game) void fetchGame();
-    if (user.currentAffiliation) void fetchPlayers();
   }, [router.query.game]);
 
   return (
@@ -215,72 +63,12 @@ const FilmRoom = () => {
             {game.two?.city} {game.two?.name}
           </Typography>
         </div>
-        {noteOpen ? (
-          <form
-            onSubmit={handleSubmit}
-            style={borderStyle}
-            className="flex w-4/5 flex-col items-center justify-center gap-4 rounded-md border-solid p-4"
-          >
-            <TextField
-              className="w-4/5"
-              name="title"
-              autoComplete="title"
-              required
-              id="title"
-              label="Title (100 characters max)"
-              onChange={handleInput}
-              value={noteDetails.title}
-              inputProps={{ maxLength: 100 }}
-            />
-            <TextField
-              className="w-full"
-              name="note"
-              autoComplete="note"
-              required
-              id="note"
-              label="Note"
-              onChange={handleInput}
-              value={noteDetails.note}
-            />
-            <Mentions
-              players={affiliatedPlayers}
-              mentions={mentions}
-              setMentions={setMentions}
-            />
-            <FormControlLabel
-              control={
-                <Checkbox
-                  checked={noteDetails.highlight}
-                  onChange={() =>
-                    setNoteDetails({
-                      ...noteDetails,
-                      highlight: !noteDetails.highlight,
-                    })
-                  }
-                  size="medium"
-                />
-              }
-              labelPlacement="end"
-              label="Highlight?"
-            />
-            <div className="flex items-center justify-center gap-2">
-              <Button type="submit" variant="contained" disabled={!isValidPlay}>
-                Submit
-              </Button>
-              <Button type="button" variant="text" onClick={() => resetPlay()}>
-                Cancel
-              </Button>
-            </div>
-          </form>
-        ) : !isClipStarted ? (
-          <Button onClick={() => startClip()} size="large">
-            Start Clip
-          </Button>
-        ) : (
-          <Button onClick={() => endClip()} size="large">
-            End Clip
-          </Button>
-        )}
+        <PlayModal
+          player={player}
+          gameId={game.id}
+          isPlayModalOpen={isPlayModalOpen}
+          setIsPlayModalOpen={setIsPlayModalOpen}
+        />
         {game.link && (
           <Youtube
             opts={{
